+++ src/componets/ModalMenu/MenuModal.jsx
@@ -7,9 +7,9 @@ import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
 import { GiCancel } from 'react-icons/gi';
 import { FaSearchLocation } from 'react-icons/fa';
 import logo from 'images/lordksix-logos_transparent.png';
-import SearchBox from './SearchBox';
+import SearchBox from 'componets/CommomComponents/SearchBox';
 import 'styles/MenuModal.css';
-import LocationBtn from './LocationBtn';
+import LocationBtn from 'componets/CommomComponents/LocationBtn';
 
 const MenuModal = () => {
   const modalRef = useRef();
